reorganise and tests
